backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.78.1
python-dotenv==1.0.0
websockets==12.0 
numpy==2.2.5
pydub==0.25.1
webrtcvad==2.0.10